Testing some rearrangements for multiprocessing
