+++ python/examples/optaa/cspp_optaa.py
@@ -6,10 +6,11 @@ from ooi_data_explorations.common import get_vocabulary, load_gc_thredds, update
 from ooi_data_explorations.uncabled.process_optaa import optaa_cspp
 
 
-# Setup needed parameters for the request, the user would need to vary these to suit their own needs and
-# sites/instruments of interest. Site, node, sensor, stream and delivery method names can be obtained from the
-# Ocean Observatories Initiative web site. The last two will set path and naming conventions to save the data
-# to the local disk
+# Setup needed parameters for the request, the user would need to vary these to
+# suit their own needs and sites/instruments of interest. Site, node, sensor,
+# stream and delivery method names can be obtained from the Ocean Observatories
+# Initiative website. The last two will set path and naming conventions to save
+# the data to the local disk
 site = 'CE02SHSP'                               # OOI Net site designator
 node = 'SP001'                                  # OOI Net node designator
 sensor = '04-OPTAAJ000'                         # OOI Net sensor designator
